Query values & parameters in background thread.
Multithreading was kind of hard to implement, but now it works.
The statuses, values and parameters of devices are queried in the
background and updated, when the server returns a result.

Now that this is working, it's possible to implement the updates
pushed by the poller.

Also, refactored more code of NicosClient to DataInputStream &
DataOutputStream.

Fixed some memory leaks after rotating (which now works without
causing errors).

Change-Id: Ic7986f69ff0da65e384f7fa646fa9374fbee0786