Filter moveables, handle rotation, save config
Steps entered by the user are now saved and restored upon reconnecting.
Steps are identified by hostname+user+devicename+coarse/fine as key.

Upon rotating the device, the application now rememberes the previously
selected device and restores the selecion.

The displayed devices are now filtered and only those with the param
abslimits are displayed (as requested by Jens Krüger). After having
finished my final exam, the moving of devices with mapping will be
readdressed (a todo remains in the code).

This is the final feature patch before finishing at least the
documentation.

Change-Id: I37e0a6f8c14f48b72dc2fd4655f0a67acbe57e24